webapp/requirements.txt
APScheduler==3.6.3
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
decorator==4.4.2
Flask==1.1.1
fsspec==0.7.1
gcsfs==0.6.1
geographiclib==1.50
geopy==1.21.0
google-api-core==1.16.0
google-auth==1.13.1
google-auth-oauthlib==0.4.1
google-cloud-bigquery==1.24.0
google-cloud-bigquery-storage==0.8.0
google-cloud-core==1.3.0
google-cloud-storage==1.27.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
grpcio==1.27.2
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
Jinja2==2.11.1
joblib==0.14.1
MarkupSafe==1.1.1
muffle==1.0.0
numpy==1.18.2
oauthlib==3.1.0
pandas==1.0.3
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22.2.post1
scipy==1.4.1
six==1.14.0
tzlocal==2.0.0
urllib3==1.25.8
Werkzeug==1.0.0
